fix(migration): make add_wasm_module migration idempotent
Use hasTable/hasColumn guards so the migration succeeds even when the
wasm_module table or proxy_host columns already exist from a previous
partial run, instead of throwing ER_TABLE_EXISTS_ERROR.

+++ backend/migrations/20260511233200_add_wasm_module.js
@@ -2,11 +2,12 @@ import { migrate as logger } from "../logger.js";
 
 const migrateName = "add_wasm_module";
 
-const up = (knex) => {
+const up = async (knex) => {
 	logger.info(`[${migrateName}] Migrating Up...`);
 
-	return knex.schema
-		.createTable("wasm_module", (table) => {
+	const tableExists = await knex.schema.hasTable("wasm_module");
+	if (!tableExists) {
+		await knex.schema.createTable("wasm_module", (table) => {
 			table.increments("id").primary();
 			table.dateTime("created_on").notNullable();
 			table.dateTime("modified_on").notNullable();
@@ -22,41 +23,48 @@ const up = (knex) => {
 			table.string("filename").notNullable();
 			table.integer("is_deleted").notNullable().defaultTo(0);
 			table.index("owner_user_id");
-		})
-		.then(() => {
-			logger.info(`[${migrateName}] 'wasm_module' table created`);
-			return knex.schema.table("proxy_host", (table) => {
-				table
-					.integer("wasm_module_id")
-					.unsigned()
-					.defaultTo(0)
-					.references("id")
-					.inTable("wasm_module")
-					.onDelete("CASCADE");
-				table.text("wasm_config").notNullable().defaultTo("");
-				table.index("wasm_module_id");
-			});
-		})
-		.then(() => {
-			logger.info(`[${migrateName}] 'wasm_module_id' and 'wasm_config' added to 'proxy_host'`);
 		});
+		logger.info(`[${migrateName}] 'wasm_module' table created`);
+	} else {
+		logger.info(`[${migrateName}] 'wasm_module' table already exists, skipping`);
+	}
+
+	const hasWasmModuleId = await knex.schema.hasColumn("proxy_host", "wasm_module_id");
+	if (!hasWasmModuleId) {
+		await knex.schema.table("proxy_host", (table) => {
+			table
+				.integer("wasm_module_id")
+				.unsigned()
+				.defaultTo(0)
+				.references("id")
+				.inTable("wasm_module")
+				.onDelete("CASCADE");
+			table.text("wasm_config").notNullable().defaultTo("");
+			table.index("wasm_module_id");
+		});
+		logger.info(`[${migrateName}] 'wasm_module_id' and 'wasm_config' added to 'proxy_host'`);
+	} else {
+		logger.info(`[${migrateName}] 'proxy_host' columns already exist, skipping`);
+	}
 };
 
-const down = (knex) => {
+const down = async (knex) => {
 	logger.info(`[${migrateName}] Migrating Down...`);
 
-	return knex.schema
-		.table("proxy_host", (table) => {
+	const hasWasmModuleId = await knex.schema.hasColumn("proxy_host", "wasm_module_id");
+	if (hasWasmModuleId) {
+		await knex.schema.table("proxy_host", (table) => {
 			table.dropColumn("wasm_module_id");
 			table.dropColumn("wasm_config");
-		})
-		.then(() => {
-			logger.info(`[${migrateName}] 'wasm_module_id' and 'wasm_config' dropped from 'proxy_host'`);
-			return knex.schema.dropTable("wasm_module");
-		})
-		.then(() => {
-			logger.info(`[${migrateName}] 'wasm_module' table dropped`);
 		});
+		logger.info(`[${migrateName}] 'wasm_module_id' and 'wasm_config' dropped from 'proxy_host'`);
+	}
+
+	const tableExists = await knex.schema.hasTable("wasm_module");
+	if (tableExists) {
+		await knex.schema.dropTable("wasm_module");
+		logger.info(`[${migrateName}] 'wasm_module' table dropped`);
+	}
 };
 
 export { up, down };
